backend/requirements.txt
Babel==2.6.0
blinker==1.4
click==6.7
emitter-io==1.0.3
Flask==1.0.2
Flask-BabelEx==0.9.3
Flask-Cors==3.0.6
Flask-JWT==0.3.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-peewee==3.0.3
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.14.5
paho-mqtt==1.3.1
pandas==0.23.3
passlib==1.7.1
pathlib==1.0.1
peewee==3.5.2
PyJWT==1.4.2
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
speaklater==1.3
Werkzeug==0.14.1
wtf-peewee==3.0.0
WTForms==2.2.1
